requirements.txt
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.30.0
openai-whisper>=20231117
streamlit>=1.24.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
pydantic>=2.0.0
python-multipart>=0.0.6
sounddevice>=0.4.6
scipy>=1.10.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
openai>=1.0.0
scikit-learn>=1.3.0
regex>=2023.6.3
tqdm>=4.66.0
nltk>=3.8.1
spacy>=3.6.0
